+++ crates/cli/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Minor
 
+- Implement `applet-rpc` command
+- Add `--serial=` to select the platform
 - Implement `platform-list` command
 - Support creating, building, and testing a Rust applet
 - Support generating a shell completion file